io: extract io_handle_accept/io_handle_connect into shared handlers.c
Both completion handlers are backend-agnostic: they run after the OS
has signalled a socket operation completed (io_uring CQE on Linux,
EVFILT_READ/WRITE on FreeBSD), and do the same work -- register the
conn_info, set TCP_NODELAY, fill in peer/local addresses, kick off
reads, re-arm the accept.  Only the submission sites
(io_request_accept_op, the connect helper) are backend-specific.

Extract the two handlers into lib/io/handlers.c, compiled on both
backends.  Remove them from accept.c (liburing-only) and connect.c
(liburing-only); remove the log-and-drop stubs from backend_kqueue.c.

No functional change on Linux.  On FreeBSD, accept_and_dispatch and
connect_and_dispatch now run real completion logic; downstream calls
into io_request_read_op and io_rpc_trans_cb still hit stubs until
PR #6 lights up the read/write handlers.
