+++ lib/io/accept.c
@@ -162,118 +162,3 @@ int io_request_accept_op(int fd, struct connection_info *ci,
 
 	return (ret < 0) ? -ret : EAGAIN;
 }
-
-int io_handle_accept(struct io_context *ic, int client_fd,
-		     struct ring_context *rc)
-{
-	char addr_str[INET6_ADDRSTRLEN];
-	uint16_t port;
-
-	// Handle new connection
-	int listen_fd = ic->ic_fd;
-
-	bool accept_resubmitted = false;
-
-	trace_io_context(ic, __func__, __LINE__);
-
-	// Always try to set up the next accept first, to ensure we don't miss connections
-	int accept_ret = io_request_accept_op(listen_fd, &ic->ic_ci, rc);
-	if (accept_ret == 0) {
-		accept_resubmitted = true;
-	} else {
-		LOG("Failed to resubmit accept request - watchdog will retry later");
-		// The watchdog will handle this
-	}
-
-	if (client_fd < 0) {
-		LOG("Accept failed with error: %s", strerror(-client_fd));
-
-		// If we haven't already resubmitted the accept, try one more time
-		if (!accept_resubmitted) {
-			LOG("Trying one more time to resubmit accept");
-			io_request_accept_op(listen_fd, &ic->ic_ci, rc);
-		}
-
-		io_context_destroy(ic);
-		return -client_fd;
-	}
-
-	int flag = 1;
-	if (setsockopt(client_fd, IPPROTO_TCP, TCP_NODELAY, &flag,
-		       sizeof(flag)) < 0) {
-		LOG("setsockopt TCP_NODELAY failed for fd=%d: %s", client_fd,
-		    strerror(errno));
-	}
-
-	// Immediately register the new client connection with ACCEPTED state
-	struct conn_info *client_conn =
-		io_conn_register(client_fd, CONN_ACCEPTED, CONN_ROLE_SERVER);
-	if (!client_conn) {
-		LOG("Failed to register client connection fd=%d", client_fd);
-		io_socket_close(client_fd, ENOMEM);
-		io_context_destroy(ic);
-		return ENOMEM;
-	}
-
-	// Get peer information
-	ic->ic_ci.ci_peer_len = sizeof(ic->ic_ci.ci_peer);
-	if (getpeername(client_fd, (struct sockaddr *)&ic->ic_ci.ci_peer,
-			&ic->ic_ci.ci_peer_len) == 0) {
-		addr_to_string(&ic->ic_ci.ci_peer, addr_str, INET6_ADDRSTRLEN,
-			       &port);
-
-		// Store peer information in connection tracking
-		memcpy(&client_conn->ci_peer, &ic->ic_ci.ci_peer,
-		       ic->ic_ci.ci_peer_len);
-		client_conn->ci_peer_len = ic->ic_ci.ci_peer_len;
-
-		TRACE("Accepted connection from %s:%d on fd=%d", addr_str, port,
-		      client_fd);
-		TRACE("ACCEPTED: fd=%d from %s:%d", client_fd, addr_str, port);
-	} else {
-		LOG("Failed to get peer information: %s", strerror(errno));
-		memset(&ic->ic_ci.ci_peer, 0, sizeof(ic->ic_ci.ci_peer));
-		ic->ic_ci.ci_peer_len = 0;
-	}
-
-	// Get local information
-	ic->ic_ci.ci_local_len = sizeof(ic->ic_ci.ci_local);
-	if (getsockname(client_fd, (struct sockaddr *)&ic->ic_ci.ci_local,
-			&ic->ic_ci.ci_local_len) == 0) {
-		addr_to_string(&ic->ic_ci.ci_local, addr_str, INET6_ADDRSTRLEN,
-			       &port);
-
-		// Store local information in connection tracking
-		memcpy(&client_conn->ci_local, &ic->ic_ci.ci_local,
-		       ic->ic_ci.ci_local_len);
-		client_conn->ci_local_len = ic->ic_ci.ci_local_len;
-	} else {
-		LOG("Failed to get local information: %s", strerror(errno));
-		memset(&ic->ic_ci.ci_local, 0, sizeof(ic->ic_ci.ci_local));
-		ic->ic_ci.ci_local_len = 0;
-	}
-
-	// Register this client in buffer state tracking
-	io_client_fd_register(client_fd);
-
-	struct conn_info *conn = io_conn_get(client_fd);
-	if (conn) {
-		// No need to check state since we're using operation counts
-		// Just ensure the state is properly set to CONNECTED initially
-		TRACE("Accepted new connection fd=%d", client_fd);
-	} else {
-		LOG("Warning: Connection fd=%d not found after registration",
-		    client_fd);
-	}
-
-	// Prepare to read from this new connection
-	io_request_read_op(client_fd, &ic->ic_ci, rc);
-
-	// Accept more connections
-	if (!accept_resubmitted) {
-		io_request_accept_op(listen_fd, &ic->ic_ci, rc);
-	}
-
-	io_context_destroy(ic);
-	return 0;
-}
