+++ lib/io/connect.c
@@ -149,78 +149,3 @@ int io_send_request(struct rpc_trans *rt)
 	return io_rpc_trans_cb(rt);
 }
 
-int io_handle_connect(struct io_context *ic, int result,
-		      struct ring_context __attribute__((unused)) * rc)
-{
-	char addr_str[INET6_ADDRSTRLEN];
-	uint16_t port;
-
-	if (result < 0) {
-		LOG("Connect failed for fd=%d: %s", ic->ic_fd,
-		    strerror(-result));
-
-		io_socket_close(ic->ic_fd, -result);
-		io_context_destroy(ic);
-		return -result;
-	}
-
-	// Try to get peer information
-	ic->ic_ci.ci_peer_len = sizeof(ic->ic_ci.ci_peer);
-	if (getpeername(ic->ic_fd, (struct sockaddr *)&ic->ic_ci.ci_peer,
-			&ic->ic_ci.ci_peer_len) != 0) {
-		LOG("Failed to get peer information: %s", strerror(errno));
-
-		// This is a critical error - the socket is not properly connected
-		io_conn_set_error(ic->ic_fd, errno);
-
-		memset(&ic->ic_ci.ci_peer, 0, sizeof(ic->ic_ci.ci_peer));
-		ic->ic_ci.ci_peer_len = 0;
-
-		io_socket_close(ic->ic_fd, errno);
-		io_context_destroy(ic);
-		return errno;
-	}
-
-	// Get local socket information
-	ic->ic_ci.ci_local_len = sizeof(ic->ic_ci.ci_local);
-	if (getsockname(ic->ic_fd, (struct sockaddr *)&ic->ic_ci.ci_local,
-			&ic->ic_ci.ci_local_len) == 0) {
-		addr_to_string(&ic->ic_ci.ci_local, addr_str, INET6_ADDRSTRLEN,
-			       &port);
-	} else {
-		LOG("Failed to get local information: %s", strerror(errno));
-		memset(&ic->ic_ci.ci_local, 0, sizeof(ic->ic_ci.ci_local));
-		ic->ic_ci.ci_local_len = 0;
-	}
-
-	TRACE("Connection established for fd=%d", ic->ic_fd);
-
-	// Update connection info with peer and local addresses
-	struct conn_info *ci = io_conn_get(ic->ic_fd);
-	if (ci) {
-		memcpy(&ci->ci_peer, &ic->ic_ci.ci_peer, ic->ic_ci.ci_peer_len);
-		ci->ci_peer_len = ic->ic_ci.ci_peer_len;
-
-		memcpy(&ci->ci_local, &ic->ic_ci.ci_local,
-		       ic->ic_ci.ci_local_len);
-		ci->ci_local_len = ic->ic_ci.ci_local_len;
-	}
-
-	// Find the corresponding RPC transaction using XID
-	struct rpc_trans *rt = io_find_request_by_xid(ic->ic_xid);
-	if (!rt) {
-		LOG("No matching rpc_trans found for XID=%u", ic->ic_xid);
-
-		io_socket_close(ic->ic_fd, ENOENT);
-		io_context_destroy(ic);
-		return ENOENT;
-	}
-
-	rt->rt_fd = ic->ic_fd;
-
-	copy_connection_info(&ic->ic_ci, &rt->rt_info.ri_ci);
-	io_context_destroy(ic);
-
-	// Now that we're connected, prepare the RPC write request
-	return io_rpc_trans_cb(rt);
-}
